git-svn-id: http://Marc-PC/svn/Full@2893 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/PnET-cohort-library/trunk/src/Cohort.cs
@@ -182,7 +182,7 @@ namespace Landis.Library.BiomassCohortsPnET
              
         {
             SubCanopyLayers = new SubCanopyLayer[IMAX];
-            for (int i = 0; i < IMAX; i++)
+            for (byte i = 0; i < IMAX; i++)
             {
                 SubCanopyLayers[i] = new SubCanopyLayer(this, i);
 
@@ -201,7 +201,7 @@ namespace Landis.Library.BiomassCohortsPnET
         public Cohort(Cohort cohort, int IMAX)
         {
             SubCanopyLayers = new SubCanopyLayer[IMAX];
-            for (int i = 0; i < IMAX; i++)
+            for (byte i = 0; i < IMAX; i++)
             {
                 SubCanopyLayers[i] = new SubCanopyLayer(this, i);
